Bugfixes improvements patch v0.5.2

+++ iRLeagueManager/ViewModels/ScheduleViewModel.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Input;
+using System.Windows;
 using System.IO;
 using Microsoft.Win32;
 
@@ -19,6 +20,7 @@ using iRLeagueManager.ViewModels.Collections;
 using System.ComponentModel;
 using System.Collections.Specialized;
 using System.Runtime.CompilerServices;
+using System.Collections;
 
 namespace iRLeagueManager.ViewModels
 {
@@ -36,7 +38,11 @@ namespace iRLeagueManager.ViewModels
         }
 
         private bool expanded = true;
-        public bool Expanded { get => expanded; set => SetValue(ref expanded, value); }
+        public bool Expanded 
+        { 
+            get => expanded; 
+            set => SetValue(ref expanded, value);
+        }
 
         public long? ScheduleId => Model?.ScheduleId;
 
@@ -63,9 +69,10 @@ namespace iRLeagueManager.ViewModels
             sessions = new ObservableModelCollection<SessionViewModel, SessionModel>(Model?.Sessions, x => x.Schedule = this);
             ((INotifyCollectionChanged)sessions).CollectionChanged += OnSessionCollectionChange;
             Sessions.UpdateSource(new SessionModel[0]);
-            AddSessionCmd = new RelayCommand(o => AddSession(), o => Model?.Sessions != null && (!Model?.IsReadOnly).GetValueOrDefault());
-            DeleteSessionsCmd = new RelayCommand(o => DeleteSessions(o), o => SelectedSession != null && (!Model?.IsReadOnly).GetValueOrDefault());
+            AddSessionCmd = new RelayCommand(async o => await AddSessionAsync(), o => Model?.Sessions != null && (!Model?.IsReadOnly).GetValueOrDefault());
+            DeleteSessionsCmd = new RelayCommand(async o => await DeleteSessionsAsync(o), o => SelectedSession != null && (!Model?.IsReadOnly).GetValueOrDefault());
             UploadFileCmd = new RelayCommand(o => UploadFile(o as SessionModel), o => false);
+            Expanded = true;
         }
 
         public ScheduleViewModel(ScheduleModel source) : this()
@@ -98,16 +105,16 @@ namespace iRLeagueManager.ViewModels
             OnPropertyChanged(nameof(SessionCount));
         }
 
-        public void AddSession()
+        public async Task AddSessionAsync()
         {
             if (Model?.Sessions == null)
                 return;
 
             var newSession = new RaceSessionModel(Model);
-            AddSession(newSession);
+            await AddSessionAsync(newSession);
         }
 
-        public async void AddSession(SessionModel session)
+        public async Task AddSessionAsync(SessionModel session)
         {
             Model.Sessions.Add(session);
 
@@ -126,32 +133,48 @@ namespace iRLeagueManager.ViewModels
             }
         }
 
-        public async void DeleteSessions(object selection)
+        public async Task DeleteSessionsAsync(object selection)
         {
             if (Model == null)
                 return;
             try
             {
+                if (MessageBox.Show("Would you really like to delete the selected Sessions?\nThis action can not be undone!", "Delete Sessions", MessageBoxButton.YesNo) != MessageBoxResult.Yes)
+                {
+                    return;
+                }
+
+                var removeList = new List<SessionModel>();
                 if (selection is SessionViewModel session) {
                     await LeagueContext.DeleteModelsAsync(session.GetSource());
-                    //Model.Sessions.Remove(session.GetSource());
+                    removeList.Add(session.GetSource());
                 }
-                else if (selection is IEnumerable<SessionViewModel> sessions)
+                else if (selection is IList multiselection)
                 {
+                    var sessions = multiselection.OfType<SessionViewModel>();
                     await LeagueContext.DeleteModelsAsync(sessions.Select(x => x.GetSource()).ToArray());
-                    foreach (var s in sessions)
-                    {
-                        //Model.Sessions.Remove(s.GetSource());
-                    }
+                    removeList.AddRange(sessions.Select(x => x.GetSource()));
                 }
                 else if (selection == null)
                 {
                     await LeagueContext.DeleteModelsAsync(SelectedSession.GetSource());
-                    //Model.Sessions.Remove(SelectedSession.GetSource());
+                    removeList.Add(selectedSession.GetSource());
                 }
 
                 IsLoading = true;
-                await LeagueContext.UpdateModelAsync(Model);
+
+                if (Model.ContainsChanges)
+                {
+                    foreach (var removeSession in removeList)
+                    {
+                        Model.Sessions.Remove(removeSession);
+                    }
+                    await LeagueContext.UpdateModelAsync(Model);
+                }
+                else
+                {
+                    await LeagueContext.GetModelAsync<ScheduleModel>(Model.ModelId, update: false, reload: true);
+                }
             }
             catch (Exception e)
             {
